disassemble: extracting InstanceCheck logic
