+++ strcs/__init__.py
@@ -8,8 +8,9 @@ from .decorator import (
     CreateArgs,
     CreatorDecorator,
 )
-from .disassemble import Field, InstanceCheck, InstanceCheckMeta, Type, TypeCache
+from .disassemble import Field, Type, TypeCache
 from .hints import resolve_types
+from .instance_check import InstanceCheck, InstanceCheckMeta
 from .meta import Meta
 from .not_specified import NotSpecified, NotSpecifiedMeta
 from .register import CreateRegister, Creator, Registerer
